requirements.txt
slack_sdk
python-dotenv
flask
